feat: allow two decimal prices and add support for shulker boxes in sorting

+++ surf-shop-core/surf-shop-core-common/src/main/kotlin/dev/slne/surf/shop/core/common/rabbit/packet/request/shop/CreateShopRequestPacket.kt
@@ -10,7 +10,7 @@ import kotlinx.serialization.Serializable
 data class CreateShopRequestPacket(
     val itemString: String,
     val storedItemCount: Int,
-    val pricePerItem: Int,
+    val pricePerItem: Double,
     val seller: SerializableUUID,
     val createdAt: SerializableOffsetDateTime
 ) : RabbitRequestPacket<SingleShopResponsePacket>()
